+++ data/beaconrestapi/src/main/java/tech/pegasys/teku/beaconrestapi/handlers/v2/validator/GetNewBlock.java
@@ -1,0 +1,89 @@
+/*
+ * Copyright 2021 ConsenSys AG.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
+ * the License. You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
+ * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations under the License.
+ */
+
+package tech.pegasys.teku.beaconrestapi.handlers.v2.validator;
+
+import static tech.pegasys.teku.beaconrestapi.RestApiConstants.GRAFFITI;
+import static tech.pegasys.teku.beaconrestapi.RestApiConstants.RANDAO_REVEAL;
+import static tech.pegasys.teku.beaconrestapi.RestApiConstants.RES_BAD_REQUEST;
+import static tech.pegasys.teku.beaconrestapi.RestApiConstants.RES_INTERNAL_ERROR;
+import static tech.pegasys.teku.beaconrestapi.RestApiConstants.RES_OK;
+import static tech.pegasys.teku.beaconrestapi.RestApiConstants.RES_SERVICE_UNAVAILABLE;
+import static tech.pegasys.teku.beaconrestapi.RestApiConstants.SERVICE_UNAVAILABLE;
+import static tech.pegasys.teku.beaconrestapi.RestApiConstants.SLOT;
+import static tech.pegasys.teku.beaconrestapi.RestApiConstants.TAG_EXPERIMENTAL;
+
+import com.fasterxml.jackson.core.JsonProcessingException;
+import io.javalin.http.Context;
+import io.javalin.http.Handler;
+import io.javalin.plugin.openapi.annotations.HttpMethod;
+import io.javalin.plugin.openapi.annotations.OpenApi;
+import io.javalin.plugin.openapi.annotations.OpenApiContent;
+import io.javalin.plugin.openapi.annotations.OpenApiParam;
+import io.javalin.plugin.openapi.annotations.OpenApiResponse;
+import tech.pegasys.teku.api.DataProvider;
+import tech.pegasys.teku.api.ValidatorDataProvider;
+import tech.pegasys.teku.api.response.v2.validator.GetNewBlockResponseV2;
+import tech.pegasys.teku.api.schema.BeaconBlock;
+import tech.pegasys.teku.provider.JsonProvider;
+
+public class GetNewBlock extends tech.pegasys.teku.beaconrestapi.handlers.v1.validator.GetNewBlock
+    implements Handler {
+  public static final String ROUTE = "/eth/v2/validator/blocks/:slot";
+
+  public GetNewBlock(final DataProvider dataProvider, final JsonProvider jsonProvider) {
+    super(dataProvider, jsonProvider);
+  }
+
+  public GetNewBlock(final ValidatorDataProvider provider, final JsonProvider jsonProvider) {
+    super(provider, jsonProvider);
+  }
+
+  @OpenApi(
+      path = ROUTE,
+      method = HttpMethod.GET,
+      summary = "Produce unsigned block",
+      tags = {TAG_EXPERIMENTAL},
+      description =
+          "Requests a beacon node to produce a valid block, which can then be signed by a validator.",
+      pathParams = {
+        @OpenApiParam(
+            name = SLOT,
+            description = "The slot for which the block should be proposed."),
+      },
+      queryParams = {
+        @OpenApiParam(
+            name = RANDAO_REVEAL,
+            description = "`BLSSignature Hex` BLS12-381 signature for the current epoch.",
+            required = true),
+        @OpenApiParam(name = GRAFFITI, description = "`Bytes32 Hex` Graffiti.")
+      },
+      responses = {
+        @OpenApiResponse(
+            status = RES_OK,
+            content = @OpenApiContent(from = GetNewBlockResponseV2.class)),
+        @OpenApiResponse(status = RES_BAD_REQUEST, description = "Invalid parameter supplied"),
+        @OpenApiResponse(status = RES_INTERNAL_ERROR),
+        @OpenApiResponse(status = RES_SERVICE_UNAVAILABLE, description = SERVICE_UNAVAILABLE)
+      })
+  @Override
+  public void handle(final Context ctx) throws Exception {
+    super.handle(ctx);
+  }
+
+  @Override
+  protected String produceResultString(final BeaconBlock block) throws JsonProcessingException {
+    return jsonProvider.objectToJSON(
+        new GetNewBlockResponseV2(provider.getMilestoneAtSlot(block.slot), block));
+  }
+}