+++ data/serializer/src/main/java/tech/pegasys/teku/api/schema/BeaconBlock.java
@@ -20,11 +20,12 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 import io.swagger.v3.oas.annotations.media.Schema;
 import java.util.Objects;
 import org.apache.tuweni.bytes.Bytes32;
+import tech.pegasys.teku.api.schema.interfaces.VersionedData;
 import tech.pegasys.teku.infrastructure.unsigned.UInt64;
 import tech.pegasys.teku.spec.Spec;
 import tech.pegasys.teku.spec.SpecVersion;
 
-public class BeaconBlock {
+public class BeaconBlock implements VersionedData {
   @Schema(type = "string", format = "uint64")
   public final UInt64 slot;